requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3
aiosignal==1.3.1 ; python_version >= '3.7'
attrs==23.2.0 ; python_version >= '3.7'
certifi==2024.2.2 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
fairlearn==0.10.0
filelock==3.13.1 ; python_version >= '3.8'
frozenlist==1.4.1 ; python_version >= '3.8'
fsspec[http]==2024.2.0 ; python_version >= '3.8'
idna==3.6 ; python_version >= '3.5'
jinja2==3.1.3 ; python_version >= '3.7'
joblib==1.3.2 ; python_version >= '3.7'
lightning-utilities==0.10.1 ; python_version >= '3.7'
markupsafe==2.1.5 ; python_version >= '3.7'
mpmath==1.3.0
multidict==6.0.5 ; python_version >= '3.7'
networkx==3.2.1 ; python_version >= '3.9'
numpy==1.26.4 ; python_version >= '3.9'
packaging==23.2 ; python_version >= '3.7'
pandas==2.2.0
pillow==10.2.0 ; python_version >= '3.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==2.2.0.post0
pytz==2024.1
pyyaml==6.0.1 ; python_version >= '3.6'
requests==2.31.0 ; python_version >= '3.7'
scikit-learn==1.4.0 ; python_version >= '3.9'
scipy==1.12.0 ; python_version >= '3.9'
setuptools==69.1.0 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sympy==1.12 ; python_version >= '3.8'
threadpoolctl==3.2.0 ; python_version >= '3.8'
timm
torch==2.2.0
torchmetrics==1.3.1 ; python_version >= '3.8'
torchvision==0.17.0
tqdm==4.66.2 ; python_version >= '3.7'
typing-extensions==4.9.0 ; python_version >= '3.8'
tzdata==2024.1 ; python_version >= '2'
urllib3==2.2.0 ; python_version >= '3.8'
yarl==1.9.4 ; python_version >= '3.7'
